Add test for Mission and navbar
