+++ src/__tests__/Navbar.test.js
@@ -1,0 +1,18 @@
+import { render } from '@testing-library/react';
+import { Provider } from 'react-redux';
+import { BrowserRouter } from 'react-router-dom';
+import Navbar from '../components/Navbar/Navbar';
+import store from '../redux/configureStore';
+
+describe('Testing component rendering properly', () => {
+  test('renders navbar component', () => {
+    const tree = render(
+      <BrowserRouter>
+        <Provider store={store}>
+          <Navbar />
+        </Provider>
+      </BrowserRouter>,
+    );
+    expect(tree).toMatchSnapshot();
+  });
+});
