+++ src/__tests__/rockets.test.js
@@ -59,7 +59,7 @@ describe('Test Rockets component', () => {
         rocket={rocketData}
       />,
     );
-    const reservedBtn = screen.getByText('CANCEL RESERVATION');
+    const reservedBtn = screen.getByText('RESERVE ROCKET');
     expect(reservedBtn).toBeInTheDocument();
   });
 });
